# Enumerate paramnames on startup