# Enumerate endpoint parameter names on startup instead of allocating a list.


- [x] You've read the [Contributor Guide](https://github.com/dotnet/aspnetcore/blob/main/CONTRIBUTING.md) and [Code of Conduct](https://github.com/dotnet/aspnetcore/blob/main/CODE-OF-CONDUCT.md).
- [ ] You've included unit or integration tests for your change, where applicable.
- [ ] You've included inline docs for your change, where applicable.
- [x] There's an open issue for the PR that you are making. If you'd like to propose a new feature or change, please open an issue to discuss the change or find an existing issue.


Summary of the changes 

## Use Enumerable instead of List to save allocation

Use of `IEnumerable<string>` when getting the endpoint parameters names from the collection, which then gets used to create the `RequestDelegateFactoryContext`

Fixes https://github.com/dotnet/aspnetcore/issues/46010