@mitchdenny 2023-01-16
Also looping in @captainsafia  since she would be interested in seeing the PR and is doing some stuff in AOT in this area.

On the surface this doesn't look like a breaking API change (good) - but in terms of performance impact in an actual application I don't think I can see the impact. One thing to point out with the benchmark is that you've got 10 route parameters - that is actually a lot when you consider even quite nested APIs such as: 

```/customers/{customer}/contacts/{contact}/addresses/{address}```

I re-ran the benchmark with 3 route parameters and the results aren't quite as dramatic in that case:

```
|                          Method |      Mean |     Error |    StdDev |   Gen0 | Allocated |
|-------------------------------- |----------:|----------:|----------:|-------:|----------:|
|              Add10StringsToList | 58.127 ns | 1.1880 ns | 1.7414 ns | 0.0278 |     120 B |
|             Add10StringsToArray | 49.984 ns | 0.8905 ns | 0.8330 ns | 0.0204 |      88 B |
|     UseEnumerableWithLinqSelect | 21.989 ns | 0.3592 ns | 0.2999 ns | 0.0167 |      72 B |
| UseEnumerableWithLocalFuncYield |  9.093 ns | 0.2014 ns | 0.1978 ns | 0.0111 |      48 B |
```

Let's imagine an application with a modest number of 3 parameter routes (say a 10):



|Scenario|Startup time contribution (10 * mean ns)|Memory contribution (10 * mean allocations)
|-|-|-|
|Without optimization|580 **nano**seconds|1200 bytes|
|With optimization|90 **nano**seconds|480 bytes|