@marafiq 2023-01-17
Thanks for taking a look at us @mitchdenny 
Three route parameters for an endpoint are a good default. One thing to note is that the code will run for every endpoint with parameters.
>but in terms of performance impact in an actual application I don't think I can see the impact

As you have seen in your benchmark run, there are slight improvements in allocation and some ns. There is not nothing else in terms of impact.